requirements.txt
pandas>=2.2.2
numpy>=1.26.4
requests>=2.32.3
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
jupyter>=1.0.0
matplotlib>=3.9.0
seaborn>=0.13.2
